requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
groq==0.11.0
supabase==2.7.4
httpx==0.27.2
sse-starlette==2.1.3
tiktoken==0.7.0
sentence-transformers==3.1.1